Fully re-implement the game in separated files.
I have also made changes to how food spawning works, it no longer relies on Clusters and instead spawns 5 food per second.

The game also now runs on deltatime entirely
